feat(article): add old petition

+++ quartz/build.ts
@@ -59,9 +59,14 @@ async function buildQuartz(argv: Argv, mut: Mutex, clientRefresh: () => void) {
 	}
 
 	const release = await mut.acquire();
-	perf.addEvent('clean');
-	await rimraf(output);
-	console.log(`Cleaned output directory \`${output}\` in ${perf.timeSince('clean')}`);
+	console.log(argv);
+	if (!argv.fastRebuild) {
+		perf.addEvent('clean');
+		await rimraf(output);
+		console.log(
+			`Cleaned ouif (argv.fastRebuild) {tput directory \`${output}\` in ${perf.timeSince('clean')}`,
+		);
+	}
 
 	perf.addEvent('glob');
 	const allFiles = await glob('**/*.*', argv.directory, cfg.configuration.ignorePatterns);
